splitting epiphyte isolation from tree isolation logic. cleaning up scratch work, old imports, etc in tree_isolation. Moving some of said code to scripts
