+++ src/tree_isolation.py
@@ -4,10 +4,7 @@ import scipy.spatial as sps
 import numpy as np
 from numpy import asarray as arr
 import pickle
-from copy import deepcopy
 import itertools
-from glob import glob
-import re
 
 from collections import defaultdict
 
@@ -18,26 +15,20 @@ from numpy import asarray as arr
 import matplotlib.pyplot as plt
 from open3d.io import read_point_cloud as read_pcd, write_point_cloud as write_pcd
 
+from geometry.zoom_and_filter import filter_to_region_pcds, zoom_pcd
 from set_config import config, log
-from reconstruction import recover_original_details, get_neighbors_kdtree, get_pcd
 from utils.math_utils import (
-    get_center,
     generate_grid
 )
-from utils.fit import kmeans,cluster_DBSCAN
-from geometry.skeletonize import extract_skeleton, extract_topology, contraction_mag_processing
-from geometry.mesh_processing import define_conn_comps, get_surface_clusters, map_density
-from geometry.point_cloud_processing import ( filter_by_norm,
+from geometry.point_cloud_processing import ( 
     clean_cloud,
-    crop, get_shape,
-    orientation_from_norms,
-    filter_by_norm,
-    get_ball_mesh,
     crop_by_percentile,
     cluster_plus
 )
-from utils.io import save,load,update,save_line_set, load_line_set
-from viz.viz_utils import iter_draw, color_continuous_map, draw, rotating_compare_gif
+# from geometry.point_cloud_processing import ( 
+# )
+from utils.io import save,load
+from viz.viz_utils import draw
 from viz.color import *
 
 def join_pcds(pcds):
@@ -467,273 +458,11 @@ def run_extend():
     # other_colored = arr(c_trunk_clusters)[~arr(refined_clusters)]
     
     other_clusters = [(idc,cluster) if idc in final_int_ids else (idc,empty_pcd) for idc,cluster in enumerate(c_trunk_clusters) ]
-    extend_seed_clusters(clusters_and_idxs = other_clusters,
-                            src_pcd=highc,
-                            file_label='other_clusters',cycles=250,save_every=20,draw_every=160, max_distance=.1)
-
-
-def color_on_percentile(pcd,
-                        val_list,
-                        pctile,
-                        comp=lambda x,y:x>y ):
-    if len(pcd.points)!= len(val_list):
-        msg = f'length of val list does not match size of pcd'
-        log.error(f'length of val list does not match size of pcd')
-        raise ValueError(msg)
-
-    color_continuous_map(pcd,val_list)
-    val = np.percentile(val_list,pctile)
-    highc_idxs = np.where(comp(val_list,val))[0]
-    highc_pcd = pcd.select_by_index(highc_idxs,invert=False)
-    lowc_pcd = pcd.select_by_index(highc_idxs,invert=True)
-    return highc_idxs, highc_pcd,lowc_pcd
-
-def get_shift(pcd, seed, iters=15, debug=False,
-              contraction = config['skeletonize']['init_contraction'],
-              attraction = config['skeletonize']['init_attraction'],):
-    """
-        Orig. run with contraction_factor 3, attraction .6, max contraction 2080 max attraction 1024
-    """
-    log.info(f'getting shift for {seed}')
-    orig = pcd
-    # log.info(f'full detail: {orig}')
-    test = orig.voxel_down_sample(voxel_size=.05)
-    # log.info(f'post vox: {test}')
-    test = test.uniform_down_sample(3)
-    # log.info(f'post vox + uni: {test}')
-    test = remove_color_pts(test,invert = True)
-    # log.info(f'post vox + uni+ rm_white: {test}')
-    ratio = len(arr(test.points))/len(arr(orig.points))
-    log.info(f'final ratio {ratio}')
-    skel_res = extract_skeleton(test, max_iter = iters, cmag_save_file=f'{contraction}_{attraction}_seed{seed}_two_voxpt05_uni3',
-                                contraction_factor=contraction,
-                                attraction_factor=attraction)
-    # contracted, total_point_shift, shift_by_step = skel_res
-    # breakpoint()
-    try:
-        topo = extract_topology(skel_res[0])
-        save_line_set(topo[0],f'{contraction}_{attraction}_seed{seed}_two_voxpt05_uni3')
-    except Exception as e:
-        log.info(f'error getting topo {e}')
-    return skel_res
-    # save(f'shifts_{file}.pkl', (total_point_shift,shift_by_step))
-
-def center_and_rotate(pcd, center=None):
-    center = pcd.get_center() if center is None else center
-    rot_90_x = np.array([[1,0,0],[0,0,-1],[0,1,0]])
-    pcd.translate(np.array([-x for x in center ]))
-    pcd.rotate(rot_90_x)
-    pcd.rotate(rot_90_x)
-    pcd.rotate(rot_90_x)
-    return center
-
-def contract(in_pcd,shift):
-    "Translates the points in the "
-    pts=arr(in_pcd.points)
-    shifted=[(pt[0]-shift[0],pt[1]-shift[1],pt[2]-shift[2]) for pt, shift in zip(pts,shift)]
-    contracted = o3d.geometry.PointCloud()
-    contracted.colors = o3d.utility.Vector3dVector(arr(in_pcd.colors))
-    contracted.points = o3d.utility.Vector3dVector(shifted)
-    return contracted
-
-def get_downsample(file = None, pcd = None, normalize = False):
-    if file: pcd = read_pcd(file)
-    log.info('Down sampling pcd')
-    voxed_down = pcd.voxel_down_sample(voxel_size=.05)
-    uni_down = voxed_down.uniform_down_sample(3)
-    clean_pcd = remove_color_pts(uni_down,invert = True)
-    if normalize: _ = center_and_rotate(clean_pcd)
-    return clean_pcd
-
-def draw_skel(pcd, 
-                seed,
-                shift=[], 
-                skeleton = None,
-                save_gif=False, 
-                out_path = None):
-    out_path = out_path or f'data/results/gif/{seed}'
-
-    clean_pcd = get_downsample(pcd=pcd, normalize=True)
-    if not skeleton:
-        skeleton = contract(clean_pcd,shift)
-    # _ = center_and_rotate(skeleton) #Rotation makes contraction harder
-    rotating_compare_gif(skeleton,clean_pcd, 
-                         point_size=4, output=out_path,
-                         save = save_gif, on_frames = 50,
-                         off_frames=50, addnl_frame_duration=.03)
-    # rotating_compare_gif(contracted,clean_pcd, point_size=4, output=out_path,save = save_gif, on_frames = 50,off_frames=50, addnl_frame_duration=.03)
-    return skeleton
-
-def draw_shift(pcd, 
-                seed,
-                shift,
-                draw_results=True, 
-                save_gif=False, 
-                out_path = None):
-    out_path = out_path or f'data/results/gif/{seed}'
-    clean_pcd = get_downsample(pcd=pcd, normalize=True)
-    c_mag = np.array([np.linalg.norm(x) for x in shift])
-    highc_idxs, highc,lowc = color_on_percentile(clean_pcd,c_mag,70)
-
-    log.info('preping contraction/coloring')
-    if draw_results:
-        draw(highc)
-        draw(lowc)
-        log.info('giffing')
-        # rotating_compare_gif(contracted,clean_pcd, output=out_path, output=out_path,save = save_gif,on_frames = 15,off_frames=3, addnl_frame_duration=.05)
-        rotating_compare_gif(highc,lowc, output=out_path, save = save_gif,on_frames = 15,off_frames=3, addnl_frame_duration=.05)
-        breakpoint()
-
-    #extrapolating point shift to the more detailed pcd
-    # # voxed_down = pcd.voxel_down_sample(voxel_size=.01)
-    # pcd, nbrs = get_neighbors_kdtree(highc_pcd,orig,k=50)
-    
-    # highc_pts = [arr(highc_pcd.points)]
-    # orig_shifts = [np.mean(c_mag[nbr_list]) for nbr_list in nbrs]
-
-
-    # # nowhite_custom =remove_color_pts(pcd, lambda x: sum(x)>2.3,invert=True)
-    # draw(highc_detail
-
-def identify_epiphytes(file, pcd, shift):
-
-    green = get_green_surfaces(pcd)
-    not_green = get_green_surfaces(pcd,True)
-    # draw(lowc_pcd)
-    draw(green)
-    draw(not_green)
-
-
-    c_mag = np.array([np.linalg.norm(x) for x in shift])
-    z_mag = np.array([x[2] for x in shift])
-
-    highc_idxs, highc, lowc = color_on_percentile(pcd,c_mag,70)
-
-    z_cutoff = np.percentile(z_mag,80)
-    log.info(f'{z_cutoff=}')
-    low_idxs = np.where(z_mag<=z_cutoff)[0]
-    lowc = clean_pcd.select_by_index(low_idxs)
-    ztrimmed_shift = shift[low_idxs]
-    ztrimmed_cmag = c_mag[low_idxs]
-    draw(lowc)
-    highc_idxs, highc, lowc = color_on_percentile(lowc,c_mag,70)
-
-
-    # color_continuous_map(test,c_mag)
-    highc_idxs = np.where(c_mag>np.percentile(c_mag,70))[0]
-    highc_pcd = test.select_by_index(highc_idxs)
-    lowc_pcd = test.select_by_index(highc_idxs,invert=True)
-    draw([lowc_pcd])
-    draw([highc_pcd])
-
-    breakpoint()
-    
-    lowc_detail = get_neighbors_kdtree(pcd,lowc_pcd)
-    draw(lowc_detail)
-    breakpoint()
-
-def get_seed(requested_seeds=[]):
-    dir = 'data/results/skio/'
-    seed_pat = re.compile('.*seed([0-9]{1,3}).*')
-    detail_files = glob('*detail*',root_dir=dir)
-    shift_files = glob('*shift*',root_dir=dir)
-    
-    seed_to_detail = {re.match(seed_pat,file).groups(1)[0]:file for file in detail_files}
-    # for seed,file in seed_to_detail.items(): get_shift(read_pcd(file),seed)
-    seed_to_shift = {re.match(seed_pat,file).groups(1)[0]:file for file in shift_files}
-    seed_to_files = [(seed,(seed_to_detail.get(seed),seed_to_shift.get(seed)))  for seed in seed_to_detail.keys() ]
-    seed_to_content = {seed:(detail,shift) for seed,(detail,shift) in seed_to_files}
-
-    # for pcd in pcd_list: draw(pcd)
-    paired_exts = [(107,108),
-                    (133,151),
-                    (189,191),
-                    (190,193),
-                    ]
-    for seed, (pcd_file, shift_file) in seed_to_content.items():
-        log.info(f'processing seed {seed}')
-        if requested_seeds==[] or int(seed) in requested_seeds:
-            # try:
-            log.info(f'processing seed {seed}')
-            log.info('loading pcd')
-            pcd = read_pcd(pcd_file)
-            log.info('loading shift')
-            shift = load(shift_file)
-            log.info('drawing shift')
-            draw_shift(pcd,shift,seed,save_gif=False)    
-            # except Exception as e:
-            #     breakpoint()
-            #     log.info(f'error with {seed},{e}')
-            breakpoint()
-
-def assess_skeletons():
-    file_types = ['lines'
-                    ,'points'
-                    #  ,'shift'
-                    ,'tpshift']
-    seed_pat = re.compile('.*seed([0-9]{1,3}).*')
-    factors_pat = re.compile('.*/([0-9]*\\.?[0-9]*_[0-9]*\\.?[0-9]*).*')
-    root_dir = 'data/results/skel_assess/'
-
-    files = defaultdict(list)
-    # for ftype in file_types:
-    points_files = glob('*_points.pkl',root_dir=root_dir)
-    detail_file = 'data/results/skio/full_ext_seed135_rf11_orig_detail.pcd'
-    pcd = read_pcd(detail_file)
-    for f in points_files:
-        if '4_2' in f:
-            factors = re.match(factors_pat,f).groups(1)[0]
-            seed = re.match(seed_pat,f).groups(1)[0]
-            base_name = f.replace('_points.pkl','')
-            tp_shift = load(f'{base_name}_tpshift.pkl',dir = root_dir)
-
-            # topo = load_line_set(base_name)
-
-            # clean_pcd = get_downsample(pcd=pcd, normalize=True)
-            # skeleton = contract(clean_pcd,tp_shift[0])
-            # draw(topo)
-            log.info(f)
-            # detail_file = glob(f'data/results/skio/{factors}_{base_name}_orig_detail.pcd')
-            # draw_skel(pcd,seed,tp_shift[0])#,skeleton=topo)
-            # draw_shift(pcd,seed,tp_shift[0])
-            breakpoint()
-
-def run_skeleton_cases():
-    seed = 135
-    file = 'data/results/skio/full_ext_seed135_rf11_orig_detail.pcd'
-    test = read_pcd(file)
-    config_list = [#(1,1), # lvl 5 monkey puzzle, has vertical lines for moss
-                   (2,1),   # monkey puzzle, not enough contraction
-                   (4,1),  #
-                   (1,1),  #
-                   (1,2),  # lvl6 monkey puzzle, has vertical lines for moss
-                   (1,4),  #
-                   (1,1),  #
-                   (4,3),  # lvl 3 monkey puzzle,
-                   (4,2),  #
-                   (2,.5), #
-                   (3,.8),  # lvl4 monkey puzzle, has vertical lines for moss
-                   (.5,2), # lvl6 monkey puzzle, has vertical lines for moss
-                   (.8,3), #
-                   (5,6),  #
-                   (6,5),] #
-    res = []
-    for cont, att in config_list:
-        try:
-            shift_res = get_shift(test,135,iters=15,contraction = cont, attraction = att)
-            res.append(shift_res)
-        except Exception as e:
-            log.info(f'error getting shift for {seed}:{e}')
+    # extend_seed_clusters(clusters_and_idxs = other_clusters,
+    #                         src_pcd=highc,
+    #                         file_label='other_clusters',cycles=250,save_every=20,draw_every=160, max_distance=.1)
 
 if __name__ =="__main__":
-    seed = 135
-    file = 'data/results/skio/full_ext_seed135_rf11_orig_detail.pcd'
-    test = read_pcd(file)
-    shift_res = get_shift(test,135,iters=15)
-    assess_skeletons()
-    breakpoint()
-    # get_seed([135])
-    breakpoint()
     # axes = o3d.geometry.create_mesh_coordinate_frame()
     # exclude_regions = {
     #         'building1' : [ (77,350,0),(100, 374,5.7)], 
@@ -801,147 +530,4 @@ if __name__ =="__main__":
     unmatched_rf_ext_idx = [idx for idx,pcd in enumerate(rf_ext_pcds) if idx not in [0 ,1 ,1 ,2,3,4 ,6,7 ,8 ,9 ,10,11,12,13,14,15,16,18] ]
     unmatched_rf_ext_pcds = [pcd for idx,pcd in enumerate(rf_ext_pcds) if idx not in [0 ,1 ,1 ,2,3,4 ,6,7 ,8 ,9 ,10,11,12,13,14,15,16,18] ]
 
-    # for seed in unmatched_seeds: 
-    #     seed_pcd,root_pcd,root_id = all_seeds[seed]
-    #     seed_pcd.paint_uniform_color([1,0,1])
-    #     draw(root_pcds+unmatched_rf_ext_pcds)
-    #     breakpoint()
-
-    # breakpoint()
-    # highc = o3d.io.read_pcd('new_collective_highc_18plus.pcd')
-    # pcd = read_pcd('partitioning_search_area_collective.pcd')
-    # highc, _ = crop_by_percentile(pcd, 17,100)
-    # unmatched_ext_pcds = arr(ext_pcds)[unmatched_clusters]
-    # unmatched_ext_pcds = [pcds_from_extend_seed_file(ext_result_file, pcd_index=i)[0] for i in unmatched_exts]
-
-    # ext_result_file = 'other_clusters_w_order_in_process.pkl'
-    # ext_result_file= 'new_seeds_w_order_in_progress.pkl'
-    # ext_pcds = pcds_from_extend_seed_file(ext_result_file)
-    # draw(unmatched_ext_pcds+all_seeds)
-    # 133,151,113,132, 191,193,194, 189]
-    
-  
-    # o3d.io.write_point_cloud(f'data/refined_clusters/k_cluster_{152}.pcd',c_tc)
-    good_ids = []
-    bad_ids = []
-    # test_seeds = [(seed,details) for seed,details in all_seeds if seed_to_exts.get(seed)]
-    # for seed, seed_details in test_seeds: draw([seed_details[1],ext_pcds[seed_to_exts.get(seed,-1)],seed_details[0]])
-    exclude = [151,107]
-    for seed, seed_details in all_seeds.items():
-        # try:
-        if seed not in exclude:
-            # seed_pcd, 
-            root_pcd, root_pcd_id = seed_details
-            prnt = False
-            reext = False
-            rf_ext_id = seed_to_exts.get(seed,-1)
-            if rf_ext_id ==-1:
-                log.info(f'ext for seed {seed} not found ')
-                bad_ids.append((seed,None,None))
-
-            # ext_pcd = ext_pcds[rf_ext_id]
-            # root_pcd.paint_uniform_color([1,0,0])
-            # # draw([root_pcd]+[ext_pcd,seed_pcd])
-
-            # detail_pcd_file = f'rf_cluster{rf_ext_id}_orig_detail.pcd'
-            # # if reext:
-            # #     # Rebuilds cluster from root_pcd to seperate tree from neighbors
-            # #     ## get original details near the new ext from the orig detail version of the old ext
-            # #     ## changes detail_pcd_file to point to the new, hopefully isolated tree
-            # #     clusters_and_idxs=[(seed,root_pcd)]#,(108,seed_to_root_map[108])]
-            # #     tree_pcds = extend_seed_clusters(clusters_and_idxs,detail_pcd,f'rf_cluster_{seed}_rebuild',cycles=700,save_every=70,draw_every=70,max_distance=.04,k=50) #
-            # #         # cycles=800,save_every=70,draw_every=70,max_distance=.05) #
-            # #     draw(tree_pcds)
-            # #     breakpoint()
-            # #     detail_pcd = read_pcd(detail_pcd_file)
-            # #     ext_pcd = tree_pcds[0]
-            # #     breakpoint()
-            # #     # Recovering original detail here, but no details for the root are available
-            # #     new_detail_pcd,pts = get_neighbors_kdtree(detail_pcd, ext_pcd)
-            # #     draw([root_pcd]+[new_detail_pcd,seed_pcd])
-            # #     breakpoint()
-            # #     # write_point_cloud(f'data/results/skio/rf_cluster{seed}_orig_detail.pcd', new_detail_pcd)
-            # #     write_point_cloud(f'rf_cluster_rebuild{seed}_orig_detail.pcd', new_detail_pcd)
-            # #     detail_pcd_file = f'rf_cluster_rebuild{seed}_orig_detail.pcd'
-            # [189,191,   193,194]
-            # First get details for :
-            # 189 matches with rf_ext_pcds[16]
-            # 193 matches with rf_ext_pcds[18]
-            partial = False
-            seed = 193
-            rf_ext_id = 18
-            root_pcd_id = seed_to_root_id[seed]
-            root_pcd = seed_to_root_map[seed]
-            seed_pcd = trunk_clusters[seed]
-            rf_pcd = rf_ext_pcds[rf_ext_id]
-            draw(root_pcd)
-            draw([seed_pcd,root_pcd,rf_pcd])
-        
-            ## The below proceeds as follows:
-            ##   0. User sets seed and rf_ext to run. Multiple provided will combine them into a single ext
-            ##   1. Manual input determines if extended seed needs to be rebuild 
-            ##      - i.e. if it is missing points
-            ##   2a. If so, the root is extened against the full set of points
-            ##   2b. If not, then the existing orig_detail file is read in 
-            ##   3. Original detail is then recovered for either:
-            #       - The new_ext (if step 2a was taken)
-            #       - Or just the existing root_pcd (if step 2b was taken)
-            #    4. All original details files are read in and joined to from the final detailed ext
-
-            #input
-            rf_ext_id = [rf_ext_id]
-            seeds = [seed]
-            #setup
-            roots = [(my_seed,seed_to_root_map[my_seed]) for my_seed in seeds]
-            root_pcd_ids = [seed_to_root_id[my_seed] for my_seed in seeds]
-            roots_pcds = [seed_to_root_map[my_seed] for my_seed in seeds]
-            rf_pcd = rf_ext_pcds[rf_ext_id]
-            draw([root_pcds])
-            draw([root_pcds])
-            breakpoint()
-            ## If partial ext, rebuild from larger source pcd
-            src_type = ''
-            if partial: # if the pcd is cutoff at a border
-                pcd = read_pcd('collective.pcd')
-                highc = crop_by_percentile(pcd,20,100)[0]
-                rf_new = extend_seed_clusters(voxed_down,highc,f'rf_cluster_root{seed}_rebuild',cycles=50,save_every=10,draw_every=10,max_distance=.05,k=300)
-                # new_new_rf = extend_seed_clusters([(seed,new_rf)],low,f'rf_cluster_{seed}_plus_down',cycles=12,save_every=3,draw_every=3,max_distance=.2,k=50) 
-                draw(rf_new)
-                get_details_pcds,src_type = rf_new,'rf_new'
-            else: # if the seed and cluster look good
-                rf_pcd = pcds_from_extend_seed_file(f'rf_cluster_root{seed}_rebuild_in_process.pkl')
-                get_details_pcds,src_type = roots_pcds,'roots_pcds'
-
-            roots_details = []
-            for root,root_id in zip(get_details_pcds,root_pcd_ids):
-                try:
-                    recover_original_details(rf_new[0],save_file=f'rf_cluster_seed{seed}_rf{11}',chunk_size=10000000)
-                except Exception as e:
-                    log.info(f'This fails but writes the file I need: {e}')
-                roots_details.append(read_pcd(f'rf_cluster_seed{seed}_root{root_id}_orig_detail'))
-            draw(roots_details)
-
-            if src_type == 'roots_pcds':
-                finals = join_pcds([roots_details,rf_pcd])[0]
-            elif src_type == 'rf_new':
-                finals = roots_details
-            draw(finals)
-
-            write_point_cloud(f'data/results/skio/full_ext_seed{'_'.join(seeds)}_rf{11}_orig_detail.pcd',finals)
-        
-        # except Exception as e:
-        #     log.info(f'error on seed {seed}: {e}')
-
-    # good_ids.extend([(seed,seed_to_exts[seed],seed_to_root_id[seed]) for seed in completed])
-
-    # save('rf_seed_ext_root_ids.pkl',good_ids)
     breakpoint()
-    # good_ids = load('rf_seed_ext_root_ids.pkl')
-    # for seed, root_ids, rf_ext_id in good_ids:
-    #     ext_pcd = ext_pcds[rf_ext_id]      
-    #     seed_pcd.paint_uniform_color([1,0,1])
-    #     good_ids.append((seed,rf_ext_id,root_pcd_id))
-    #     recover_original_details(root_details,save_file=f'rf_root_cluster_{seed}')
-    #     root_details = f'rf_root_cluster_{seed}_orig_detail.pcd'
-    #     final = join_pcds([root_details, ext_pcd])   
-    #     o3d.io.write_point_cloud(root_details,final)
